V5.3.3
Delete/ archive requires confirmations, buttons are red.

+++ frontend/my-app/src/Admins/PredefinedProducts.jsx
@@ -35,6 +35,9 @@ let Products=()=>{
     }
 
     let deleteProduct=(productID)=>{
+
+        if(window.confirm("Stargeti produsul?") === false) return false
+
         fetch(`http://localhost:3000/products/${productID}`, {
             method:"DELETE",
             headers: { 'Content-Type': 'application/json'
@@ -95,7 +98,7 @@ let Products=()=>{
                                         <td>{element.pp_tax}%</td>
                                         <td>{element.pp_description}</td>
                                         <td className="table-actions-container">                            
-                                            <button title="Sterge produs" onClick={()=>{deleteProduct(element.id)}}><div class="inner-button-content"><span class="material-icons-outlined">delete</span></div></button>
+                                            <button title="Sterge produs" onClick={()=>{deleteProduct(element.id)}}><div class="inner-button-content"><span class="material-icons-outlined text-danger">delete</span></div></button>
                                         </td>
                                     </tr>
                                 )):"Nu exista produse"
